# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MMO Battle Game API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /player:
    get:
      operationId: getPlayer
      tags: [player]
      summary: Get player info
      responses:
        "200":
          description: Player data
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
        "404":
          description: No player yet
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      operationId: createPlayer
      tags: [player]
      summary: Create or reset a player
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePlayerRequest"
      responses:
        "200":
          description: Player created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Player"
...
  /inventory/{id}/sell:
    post:
      operationId: sellItem
      tags: [inventory]
      summary: Sell an item for gold
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Item sold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellResponse"
        "400":
          description: Cannot sell equipped item
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Item not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /inventory/sell-all:
    post:
      operationId: sellAllItems
      tags: [inventory]
      summary: Sell all unequipped items of a given rarity (or all unequipped if no rarity)
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SellAllRequest"
      responses:
        "200":
          description: Items sold
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SellAllResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    CreatePlayerRequest:
      type: object
      properties:
        name:
          type: string
      required:
        - name
    Player:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        level:
          type: integer
        xp:
          type: integer
        xpToNextLevel:
          type: integer
        hp:
          type: integer
        maxHp:
          type: integer
        attack:
          type: integer
        defense:
          type: integer
        gold:
          type: integer
        monstersDefeated:
          type: integer
        enchantingStones:
          type: integer
        meleeSkillLevel:
          type: integer
        meleeSkillXp:
          type: integer
        meleeSkillXpToNext:
          type: integer
        defenseSkillLevel:
          type: integer
        defenseSkillXp:
          type: integer
        defenseSkillXpToNext:
          type: integer
        ascensionLevel:
          type: integer
      required:
        - id
        - name
        - level
        - xp
        - xpToNextLevel
        - hp
        - maxHp
        - attack
        - defense
        - gold
        - monstersDefeated
        - enchantingStones
        - meleeSkillLevel
        - meleeSkillXp
        - meleeSkillXpToNext
        - defenseSkillLevel
        - defenseSkillXp
        - defenseSkillXpToNext
        - ascensionLevel
...
    SellResponse:
      type: object
      properties:
        goldGained:
          type: integer
        player:
          $ref: "#/components/schemas/Player"
      required:
        - goldGained
        - player
    SellAllRequest:
      type: object
      properties:
        rarity:
          type: string
          enum: [Common, Uncommon, Rare, Epic, Legendary, Mythic, Divine, Abyssal, Transcendent, Cosmic, Eternal]
    SellAllResponse:
      type: object
      properties:
        count:
          type: integer
        goldGained:
          type: integer
        player:
          $ref: "#/components/schemas/Player"
      required:
        - count
        - goldGained
        - player